refactor(controller): move type PluginSchemaHelper and fn fill()
